requirements.txt
Flask==3.0.3
flask-cors==4.0.0
gunicorn==22.0.0
psycopg[binary,pool]==3.2.10
python-dotenv==1.0.1
